requirements.txt
Django>=3.2.14,<4.0
boto3==1.24.36
dj_database_url==0.5.0
django-cleanup==6.0.0
djangorestframework==3.14.0
django-storages==1.13.1
music-metadata-territories==22.6
Markdown==3.4.1
Pillow==9.2.0
psycopg2-binary==2.9.3
waitress==2.1.2
whitenoise==6.2.0
